Fixed datatable error when filtering a Datetime or other complex column
When filtering a column whose qname is an expression instead of a plain
field (es: 'qname' => 'CAST("pr"."dataora" AS DATE)',) a mysql error was
thrown due to the column name present in the HAVING not being found in
the SELECT (not recognized as casted).
Now it is possible to specify a plain colum name with 'cname' and, if
qname is recognized as an expression, cname will be added to select to
make the query valid.